+++ src/core/rogo/question_classes.py
@@ -4,234 +4,24 @@
 
 from __future__ import annotations
 
-from dataclasses import dataclass
-from enum import Enum
-from functools import total_ordering
-from typing import Any, Protocol
-
-from ..accido.misc import EndingComponents
-
-
-@total_ordering
-class QuestionClasses(Enum):
-    """The classes of questions that can be asked."""
-
-    TYPEIN_ENGTOLAT = "TypeInEngToLatQuestion"
-    TYPEIN_LATTOENG = "TypeInLatToEngQuestion"
-    PARSEWORD_LATTOCOMP = "ParseWordLatToCompQuestion"
-    PARSEWORD_COMPTOLAT = "ParseWordCompToLatQuestion"
-    PRINCIPAL_PARTS = "PrincipalPartsQuestion"
-    MULTIPLECHOICE_ENGTOLAT = "MultipleChoiceEngToLatQuestion"
-    MULTIPLECHOICE_LATTOENG = "MultipleChoiceLatToEngQuestion"
-
-    def __lt__(self, other: object) -> bool:
-        if not isinstance(other, QuestionClasses):
-            return NotImplemented
-        return self.value < other.value
-
-
-class Question(Protocol):
-    """A protocol for questions."""
-
-    prompt: Any
-
-    def check(self, response: Any) -> bool:
-        """Check if the response is correct."""
-        ...
-
-
-@dataclass
-class _MultiAnswerQuestion[T]:
-    """Generic class for questions."""
-
-    main_answer: T
-    answers: set[T]
-
-    def check(self, response: T) -> bool:
-        """Check if the response is correct.
-
-        Parameters
-        ----------
-        response : T
-            The response to the question.
-
-        Returns
-        -------
-        bool
-            ``True`` if the response is correct, ``False`` otherwise.
-        """
-        return response in self.answers
-
-
-@dataclass
-class _MultipleChoiceQuestion:
-    """Generic class for multiple choice questions."""
-
-    prompt: str
-    answer: str
-    choices: tuple[str, ...]
-
-    def check(self, response: str) -> bool:
-        """Check if the given answer is correct.
-
-        Parameters
-        ----------
-        response : str
-            The answer to check.
-
-        Returns
-        -------
-        bool
-            ``True`` if the given answer is correct, ``False`` otherwise.
-        """
-        return response == self.answer
-
-
-@dataclass
-class TypeInEngToLatQuestion(_MultiAnswerQuestion[str]):
-    """A question that asks for the Latin translation of an English word.
-
-    For example, "I search" (answer: "quaero")
-
-    Attributes
-    ----------
-    prompt : str
-        The prompt for the question (in English).
-    main_answer : str
-        The best answer to the question (in Latin).
-    answers : set[str]
-        The possible answers to the question (in Latin).
-    """
-
-    prompt: str
-
-
-@dataclass
-class TypeInLatToEngQuestion(_MultiAnswerQuestion[str]):
-    """A question that asks for the English translation of a Latin word.
-
-    For example, "quaero" (answer: "I search")
-
-    Attributes
-    ----------
-    prompt : str
-        The prompt for the question (in Latin).
-    main_answer : str
-        The best answer to the question (in English).
-    answers : set[str]
-        The possible answers to the question (in English).
-    """
-
-    prompt: str
-
-
-@dataclass
-class ParseWordCompToLatQuestion(_MultiAnswerQuestion[str]):
-    """A question that asks for a Latin word given the grammatical
-    components of the word.
-
-    For example:
-    present active indicative 2nd person plural of "quaero"
-    (answer: "quaeratis").
-
-    Attributes
-    ----------
-    prompt : str
-        The prompt for the question (the dictionary entry).
-    components : EndingComponents
-        The grammatical components of the word.
-    main_answer : str
-        The best answer to the question.
-    answers : set[str]
-        The possible answers to the question.
-    """  # noqa: D205
-
-    prompt: str
-    components: EndingComponents
-
-
-@dataclass
-class ParseWordLatToCompQuestion(_MultiAnswerQuestion[EndingComponents]):
-    """A question that asks for the grammatical components of a Latin
-    word, given the word.
-
-    For example:
-    Parse "quaeratis" (hear: quaero, quaerere, quaesivi, quaesitus)
-    (answer: "present active indicative 2nd person plural").
-
-    Attributes
-    ----------
-    prompt : str
-        The prompt for the question.
-    dictionary_entry : str
-        The dictionary entry for the word.
-    main_answer : EndingComponents
-        The best answer to the question.
-    answers : set[EndingComponents]
-        The possible answers to the question.
-    """  # noqa: D205
-
-    prompt: str
-    dictionary_entry: str
-
-
-@dataclass
-class PrincipalPartsQuestion:
-    """A question that asks for the principal parts of a Latin verb.
-
-    Attributes
-    ----------
-    prompt : str
-        The prompt of the question.
-    principal_parts : tuple[str, ...]
-        The answer of the question (the principal parts).
-    """
-
-    prompt: str
-    principal_parts: tuple[str, ...]
-
-    def check(self, response: tuple[str, ...]) -> bool:
-        """Check if the given principal parts are correct.
-
-        Parameters
-        ----------
-        response : tuple[str, ...]
-            The principal parts to check.
-
-        Returns
-        -------
-        bool
-            ``True`` if the given principal parts are correct,
-            ``False`` otherwise.
-        """
-        return response == self.principal_parts
-
-
-@dataclass
-class MultipleChoiceEngToLatQuestion(_MultipleChoiceQuestion):
-    """An English to Latin multiple choice question.
-
-    Attributes
-    ----------
-    prompt : str
-        The prompt of the question (in English).
-    answer : str
-        The answer of the question (in Latin).
-    choices : tuple[str, ...]
-        The choices of the question (including the answer).
-    """
-
-
-@dataclass
-class MultipleChoiceLatToEngQuestion(_MultipleChoiceQuestion):
-    """A Latin to English multiple choice question.
-
-    Attributes
-    ----------
-    prompt : str
-        The prompt of the question (in Latin).
-    answer : str
-        The answer of the question (in English).
-    choices : tuple[str, ...]
-        The choices of the question (including the answer).
-    """
+from ._base import (
+    MultiAnswerQuestion as MultiAnswerQuestion,
+    MultipleChoiceQuestion as MultipleChoiceQuestion,
+    Question as Question,
+    QuestionClasses as QuestionClasses,
+)
+from ._multiplechoice_engtolat import (
+    MultipleChoiceEngToLatQuestion as MultipleChoiceEngToLatQuestion,
+)
+from ._multiplechoice_lattoeng import (
+    MultipleChoiceLatToEngQuestion as MultipleChoiceLatToEngQuestion,
+)
+from ._parseword_comptolat import (
+    ParseWordCompToLatQuestion as ParseWordCompToLatQuestion,
+)
+from ._parseword_lattocomp import (
+    ParseWordLatToCompQuestion as ParseWordLatToCompQuestion,
+)
+from ._principal_parts import PrincipalPartsQuestion as PrincipalPartsQuestion
+from ._typein_engtolat import TypeInEngToLatQuestion as TypeInEngToLatQuestion
+from ._typein_lattoeng import TypeInLatToEngQuestion as TypeInLatToEngQuestion
